requirements.txt
torch>=1.11.0
torchvision>=0.12.0
timm==0.5.4
open-clip-torch>=2.0.0
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
tqdm>=4.62.0
Pillow>=8.3.0
ftfy>=6.0.0
regex>=2021.0.0
matplotlib>=3.4.0
requests>=2.26.0
